requirements.txt
telethon
aiogram
python-dotenv
aiosqlite
apscheduler
pydantic
